Move code into multiple files
