+++ unity/Assets/Maroon/reusableScripts/Physics/Motion/IIntegrator.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Maroon.Physics.Motion;
+
+public interface IIntegrator
+{
+    public State Integrate(State initial, float t, float dt);
+}
