+++ unity/Assets/Maroon/reusableScripts/Physics/Motion/IIntegrator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de533f32081726f4db4bd257400bc588
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
