+++ unity/Assets/Maroon/reusableScripts/Physics/Motion/Integrators.cs
@@ -1,0 +1,81 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Maroon.Physics.Motion
+{
+    /// <summary>
+    /// Explicit Euler Integrator
+    /// </summary>
+    public class ExplicitEuler : IIntegrator
+    {
+        public State Integrate(State s, float t, float dt)
+        {
+            s.position += s.velocity * dt;
+            s.velocity += s.acceleration(t + dt) * dt;
+
+            return s;
+        }
+    }
+
+
+    /// <summary>
+    /// Semi-Implicit Euler Integrator
+    /// </summary>
+    public class SemiImplicitEuler : IIntegrator
+    {
+        public State Integrate(State s, float t, float dt)
+        {
+            s.velocity += s.acceleration(t + dt) * dt;
+            s.position += s.velocity * dt;
+
+            return s;
+        }
+    }
+
+
+    /// <summary>
+    /// 4th Order Runge-Kutta Integrator
+    /// </summary>
+    public class RungeKutta4 : IIntegrator
+    {
+        public State Integrate(State s, float t, float dt)
+        {
+            var k1 = Evaluate(s, t, 0.0f, new Derivative());
+            var k2 = Evaluate(s, t, dt * 0.5f, k1);
+            var k3 = Evaluate(s, t, dt * 0.5f, k2);
+            var k4 = Evaluate(s, t, dt, k3);
+
+            var dxdt = 1.0f / 6.0f * (k1.dx + 2.0f * k2.dx + 2.0f * k3.dx + k4.dx);
+            var dvdt = 1.0f / 6.0f * (k1.dv + 2.0f * k2.dv + 2.0f * k3.dv + k4.dv);
+
+            s.position += dxdt * dt;
+            s.velocity += dvdt * dt;
+
+            return s;
+        }
+
+        private Derivative Evaluate(State s, float t, float dt, Derivative d)
+        {
+            return new Derivative(s.velocity + d.dv * dt, s.acceleration(t + dt));
+        }
+
+        private class Derivative
+        {
+            public Vector3 dx; // velocity
+            public Vector3 dv; // acceleration
+
+            public Derivative(Vector3 dx, Vector3 dv)
+            {
+                this.dx = dx;
+                this.dv = dv;
+            }
+
+            public Derivative()
+            {
+
+            }
+        }
+    }
+
+}
