+++ unity/Assets/Maroon/reusableScripts/Physics/Motion/Integrators.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19eb4a9fa59409544ab94279660e1c93
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
