+++ unity/Assets/Maroon/reusableScripts/Physics/Motion/Solver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8070bfbc4f02211488129da9ce84d8ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
